Remove all traces of the existing MSVC build system
This is more of a PoC, not intended to be merged.
It should prove that the MSVC build system removal
does not break the autotools based build.
But before the actual merge it should be rebased on
top of the patch that introduces the CMake buildsystem
which is intended to replace this functionality.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>