configure: Try to detect LZO with pkg-config
On most systems this should work just fine.

Also remove the handling to autodetect -llzo. This
seems to be very uncommon now (we do not have examples
of it in our build farm). If you have a system that
needs -llzo just specify LZO_LIBS or configure pkg-config
correctly.

v2:
 - simplify code by removing -llzo special handling

Change-Id: I1c038dc4ec80d3499582d81eee61fee74f26e693
Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>